python/requirements.txt
faiss-cpu==1.8.0
numpy==1.26.4
scikit_learn==1.5.1
datasets==2.21.0